refactor(server): extract track_usage to services/analytics.py
Task C of docs/current/server-py-module-split.md. Pure extraction,
verbatim move verified byte-identical against git history. New module
owns its own sqlite connection, structlog logger, and ANALYTICS_DB_PATH
import. Removed now-dead sqlite3 and datetime imports from server.py
(track_usage was their only consumer). lifespan's two call sites are
unchanged; only the import moved. Functional check inserts and reads a
real usage_events row from the new location.
